리스트 및 상세보기

+++ src/main/java/com/example/demo/domain/Question/dto/QuizDto.java
@@ -1,17 +1,17 @@
 package com.example.demo.domain.Question.dto;
 
 public class QuizDto {
-  private int quizid;
+  private Long quizid;
 
-  public QuizDto(int quizid) {
+  public QuizDto(Long quizid) {
     this.quizid = quizid;
   }
 
-  public int getQuizid() {
+  public Long getQuizid() {
     return quizid;
   }
 
-  public void setQuizid(int quizid) {
+  public void setQuizid(Long quizid) {
     this.quizid = quizid;
   }
 }
